// src/app/models/Application.ts
export class Application {
    id: string;
    first_name: string;
    last_name: string;
    dob: string;
    email: string;
    phone: number;
    address: string;
    driving_license: string;
    employer_name: string;
    employer_address: string;
    job_title: string;
    annual_salary: number;
    monthly_inhand_salary: number;
    bank_name: string;
    account_number: string;
    accout_type: string;
    status: number;
    status_name: string;
    status_changed_by: string;
    read: boolean;
    assigned: boolean;
    customer_id: number;
    delete: boolean;
    modified_by: string;
    assigned_to: string;
    assigned_name: string;
}